+++ src/engine/scene/scene_base.h
@@ -46,11 +46,11 @@ public:
     SceneBase& operator=(SceneBase&&) = delete;
 
     // 核心循环方法
-    virtual void init();                   ///< @brief 初始化场景。
-    virtual void update(double deltaTime); ///< @brief 更新场景。
-    virtual void render();                 ///< @brief 渲染场景。
-    virtual void handleInput();            ///< @brief 处理输入。
-    virtual void clean();                  ///< @brief 清理场景。
+    virtual void init();                  ///< @brief 初始化场景。
+    virtual void update(float deltaTime); ///< @brief 更新场景。
+    virtual void render();                ///< @brief 渲染场景。
+    virtual void handleInput();           ///< @brief 处理输入。
+    virtual void clean();                 ///< @brief 清理场景。
 
     /// @brief 直接向场景中添加一个游戏对象。（初始化时可用，游戏进行中不安全） （&&表示右值引用，与std::move搭配使用，避免拷贝）
     virtual void addGameObject(std::unique_ptr<engine::object::GameObject>&& gameObject);
